Fix up how scope is tracked
As adding code to deal with #52, discovered that when an ast has a rep assigned to it - we don't visit it. Which makes sense. But when we don't visit it, sometimes the current scope is now wrong! So - where to recover the scope from? Decided to add the scope to every single representation.

And ran into another sort-of interesting thing: If a loop variable gets a scope, then the scope for that loop variable needs to change when it is inside a Where!