Shift to useing cpp_sequences and values
The heuristics I was using to tell them apart were not working well - the .First() stuff forced me to re-arctect that portion of the code.

This isn't perfect, but it is a lot better! And it "makes" sense now. There are still likley problems with sequencs of sequences of sequences, but for now this is working well.